+++ packages/cli/src/commands/build-all-private/index.ts
@@ -1,10 +1,9 @@
-import chalk from 'chalk';
 import { Command } from 'commander';
+
+import { red } from '../../utils/index.js';
 
 export default new Command()
   .name('build-all')
-  .description(
-    `Builds all of the rapidstack packages. ${chalk.red('[Internal]')}`
-  )
+  .description(`Builds all of the rapidstack packages. ${red('[Internal]')}`)
   .option('-d, --debug', 'output extra debug logging')
   .action(() => console.log('hello!'));
