+++ packages/cli/src/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './constants.js';
 export * from './shell.js';
 export * from './logger.js';
